backend_fastapi/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
duckduckgo-search
requests
python-dotenv
